+++ apps/CHANGELOG.md
@@ -1,3 +1,28 @@
+## 20250425-0.2.2-alpha
+
+
+### Bug Fixes
+
+* assignment without positions ([4665a68](https://github.com/sutanto62/lilium/commit/4665a682a36469f2f7148d3d595d495eb4eabdda))
+* auth provider error ([01c315d](https://github.com/sutanto62/lilium/commit/01c315dd9899f8f851b0967a552ea8d430647158))
+* error 500 ([a152509](https://github.com/sutanto62/lilium/commit/a1525097dd37dc337a8e8b22205b25c716f5c349))
+* error svelte-tags-input ([fceab41](https://github.com/sutanto62/lilium/commit/fceab41277c5b060dd539ef94bd96137beb92fdc))
+* event created without position ([#7](https://github.com/sutanto62/lilium/issues/7)) ([#8](https://github.com/sutanto62/lilium/issues/8)) ([52de11b](https://github.com/sutanto62/lilium/commit/52de11bb92b41d46a45c581c8682e82e690d54e7))
+* favicon.ico error ([a9c58ac](https://github.com/sutanto62/lilium/commit/a9c58ac7b6e045b21b2a7d6d408e7a2d1d4845d8))
+* past events design ([1064a8d](https://github.com/sutanto62/lilium/commit/1064a8d2fe869ea809bdcb081713076ee7eea603))
+* pic layout ([c20029f](https://github.com/sutanto62/lilium/commit/c20029f3ecce2f0e6feb661b111ea282b84d52af))
+* posthog setup ([65ff036](https://github.com/sutanto62/lilium/commit/65ff036471ed2b545a6f52451fab97a401b142d5))
+* round robin with PPG positions first ([3be6316](https://github.com/sutanto62/lilium/commit/3be6316867bb22b738be3b942422ba23895927e1))
+* sso button ([#9](https://github.com/sutanto62/lilium/issues/9)) ([db9a740](https://github.com/sutanto62/lilium/commit/db9a74010f73fb142399a6aeb90ab342f1d32cd4))
+
+
+### Features
+
+* download usher list ([e7fc977](https://github.com/sutanto62/lilium/commit/e7fc977731b92f766a08cd63f0814c132464df6c))
+* event and usher management ([284d99a](https://github.com/sutanto62/lilium/commit/284d99a4aaec158144836e510da4086b10e58181))
+
+
+
 ## 0.0.1 (2025-03-21)